Move all the "sun" logic into a module
