+++ src/lib.rs
@@ -1,10 +1,6 @@
 //! Loads and renders a glTF file as a scene.
 
-use bevy::{
-    pbr::{CascadeShadowConfigBuilder, DirectionalLightShadowMap},
-    prelude::*,
-};
-use std::f32::consts::*;
+use bevy::prelude::*;
 use tilemap::TileMap;
 #[cfg(all(feature = "xr", not(target_os = "macos")))]
 use xr::XRPlugin;
@@ -14,6 +10,7 @@ use bevy_flycam::prelude::*;
 use bevy_oxr::DefaultXrPlugins;
 
 mod flycam;
+mod sun;
 mod tilemap;
 #[cfg(all(feature = "xr", not(target_os = "macos")))]
 mod xr;
@@ -27,15 +24,14 @@ pub fn main() {
     } else {
         app.add_plugins(DefaultPlugins);
     }
-    app.insert_resource(DirectionalLightShadowMap { size: 4096 })
-        .insert_resource(Msaa::Sample4) // Msaa::Sample4  Msaa::default()   -- Todo: tut nichts?
+    sun::init(&mut app);
+    app.insert_resource(Msaa::Sample4) // Msaa::Sample4  Msaa::default()   -- Todo: tut nichts?
         .add_plugins(bevy::diagnostic::LogDiagnosticsPlugin::default())
         .add_plugins(bevy::diagnostic::FrameTimeDiagnosticsPlugin)
         .add_systems(Startup, (setup, flycam::setup))
         .add_systems(
             Update,
             (
-                animate_light_direction,
                 update_active_tile_zone,
                 tilemap::update,
                 flycam::move_shape_with_camera,
@@ -46,25 +42,6 @@ pub fn main() {
 }
 
 fn setup(mut commands: Commands, mut meshes: ResMut<Assets<Mesh>>) {
-    commands.spawn(DirectionalLightBundle {
-        directional_light: DirectionalLight {
-            illuminance: 15000., // lux  https://docs.rs/bevy/latest/bevy/pbr/struct.DirectionalLight.html
-            shadows_enabled: true,
-            ..default()
-        },
-        // This is a relatively small scene, so use tighter shadow
-        // cascade bounds than the default for better quality.
-        // We also adjusted the shadow map to be larger since we're
-        // only using a single cascade.
-        cascade_shadow_config: CascadeShadowConfigBuilder {
-            num_cascades: 1,
-            //maximum_distance: 1.6,
-            ..default()
-        }
-        .into(),
-        ..default()
-    });
-
     commands.spawn(TileMap::new(&mut meshes));
 }
 
@@ -81,18 +58,4 @@ fn update_active_tile_zone(mut commands: Commands, mut tilemap: Query<&mut TileM
     tilemap.load(&mut commands, 17431, 11369);
     tilemap.load(&mut commands, 17431, 11370);
     tilemap.load(&mut commands, 17431, 11371);
-}
-
-fn animate_light_direction(
-    time: Res<Time>,
-    mut query: Query<&mut Transform, With<DirectionalLight>>,
-) {
-    for mut transform in &mut query {
-        transform.rotation = Quat::from_euler(
-            EulerRot::ZYX,
-            0.0,
-            time.elapsed_seconds() * PI / 100.0,
-            -FRAC_PI_4,
-        );
-    }
 }
