+++ src/sun.rs
@@ -1,0 +1,48 @@
+//! Everything related to the global light and shadow logic
+
+use bevy::{
+    pbr::{CascadeShadowConfigBuilder, DirectionalLightShadowMap},
+    prelude::*,
+};
+use std::f32::consts::*;
+
+pub fn init(app: &mut App) {
+    app.insert_resource(DirectionalLightShadowMap { size: 4096 })
+        .add_systems(Startup, setup)
+        .add_systems(Update, (animate_light_direction,));
+}
+
+fn animate_light_direction(
+    time: Res<Time>,
+    mut query: Query<&mut Transform, With<DirectionalLight>>,
+) {
+    for mut transform in &mut query {
+        transform.rotation = Quat::from_euler(
+            EulerRot::ZYX,
+            0.0,
+            time.elapsed_seconds() * PI / 100.0,
+            -FRAC_PI_4,
+        );
+    }
+}
+
+fn setup(mut commands: Commands) {
+    commands.spawn(DirectionalLightBundle {
+        directional_light: DirectionalLight {
+            illuminance: 15000., // lux  https://docs.rs/bevy/latest/bevy/pbr/struct.DirectionalLight.html
+            shadows_enabled: true,
+            ..default()
+        },
+        // This is a relatively small scene, so use tighter shadow
+        // cascade bounds than the default for better quality.
+        // We also adjusted the shadow map to be larger since we're
+        // only using a single cascade.
+        cascade_shadow_config: CascadeShadowConfigBuilder {
+            num_cascades: 1,
+            //maximum_distance: 1.6,
+            ..default()
+        }
+        .into(),
+        ..default()
+    });
+}
